requirements.txt
langchain==0.2.6
openai==1.35.10
anthropic==0.30.1
langchain-openai==0.1.14
langchain-anthropic==0.1.19
fastapi==0.111.0
colorama~=0.4.6
langgraph==0.1.5
python-dotenv==1.0.1
pydantic==2.5.3
langchain-core==0.2.11
langchain_experimental==0.0.62